backend/helpers/llm_helper.py
import os
from groq import Groq
from dotenv import load_dotenv

load_dotenv()

def get_groq_response(prompt):
    """
    Gets a response from the Groq LLM.
    """
    try:
        client = Groq(api_key=os.environ.get("GROQ_API_KEY"))
        chat_completion = client.chat.completions.create(
            messages=[
                {
                    "role": "user",
                    "content": prompt,
                }
            ],
            model="llama3-70b-8192",
        )
        return chat_completion.choices[0].message.content
    except Exception as e:
        print(f"An error occurred: {e}")
        return None